requirements.txt
beautifulsoup4==4.8.2
python-dotenv==0.12.0
python-telegram-bot==12.4.2
requests==2.23.0
